@@ -1,18 +1,34 @@
 module github.com/elastic/go-sysinfo
 
-go 1.17
+go 1.18
 
 require (
+	github.com/docker/docker v24.0.7+incompatible
 	github.com/elastic/go-windows v1.0.0
 	github.com/joeshaw/multierror v0.0.0-20140124173710-69b34d4ec901
-	github.com/prometheus/procfs v0.0.0-20190425082905-87a4384529e0
-	github.com/stretchr/testify v1.3.0
-	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a
+	github.com/prometheus/procfs v0.8.0
+	github.com/stretchr/testify v1.7.0
+	golang.org/x/sys v0.13.0
 	howett.net/plist v0.0.0-20181124034731-591f970eefbb
 )
 
 require (
-	github.com/davecgh/go-spew v1.1.0 // indirect
-	github.com/pkg/errors v0.8.1 // indirect
+	github.com/Microsoft/go-winio v0.6.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/docker/distribution v2.8.2+incompatible // indirect
+	github.com/docker/go-connections v0.4.0 // indirect
+	github.com/docker/go-units v0.5.0 // indirect
+	github.com/gogo/protobuf v1.3.2 // indirect
+	github.com/moby/term v0.0.0-20221205130635-1aeaba878587 // indirect
+	github.com/morikuni/aec v1.0.0 // indirect
+	github.com/opencontainers/go-digest v1.0.0 // indirect
+	github.com/opencontainers/image-spec v1.0.2 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/mod v0.6.0-dev.0.20220419223038-86c51ed26bb4 // indirect
+	golang.org/x/net v0.17.0 // indirect
+	golang.org/x/time v0.3.0 // indirect
+	golang.org/x/tools v0.1.12 // indirect
+	gopkg.in/yaml.v3 v3.0.0 // indirect
+	gotest.tools/v3 v3.4.0 // indirect
 )
