@@ -15,18 +15,21 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//go:build (amd64 && cgo) || (arm64 && cgo)
-// +build amd64,cgo arm64,cgo
+//go:build amd64 || arm64
 
 package darwin
 
-import "fmt"
+import (
+	"fmt"
+
+	"golang.org/x/sys/unix"
+)
 
 const hwMemsizeMIB = "hw.memsize"
 
 func MemTotal() (uint64, error) {
-	var size uint64
-	if err := sysctlByName(hwMemsizeMIB, &size); err != nil {
+	size, err := unix.SysctlUint64(hwMemsizeMIB)
+	if err != nil {
 		return 0, fmt.Errorf("failed to get mem total: %w", err)
 	}
 
