@@ -45,7 +45,7 @@ func parseVMStat(content []byte) (*types.VMStatInfo, error) {
 	var vmStat types.VMStatInfo
 	refValues := reflect.ValueOf(&vmStat).Elem()
 
-	err := parseKeyValue(content, " ", func(key, value []byte) error {
+	err := parseKeyValue(content, ' ', func(key, value []byte) error {
 		// turn our []byte value into an int
 		val, err := parseBytesOrNumber(value)
 		if err != nil {
