@@ -20,10 +20,15 @@ package types
 import "time"
 
 // Host is the interface that wraps methods for returning Host stats
+// It may return partial information if the provider
+// implementation is unable to collect all of the necessary data.
 type Host interface {
 	CPUTimer
 	Info() HostInfo
 	Memory() (*HostMemoryInfo, error)
+
+	// FQDN returns the fully-qualified domain name of the host, lowercased.
+	FQDN() (string, error)
 }
 
 // NetworkCounters represents network stats from /proc/net
@@ -64,7 +69,7 @@ type HostInfo struct {
 	Architecture      string    `json:"architecture"`            // Hardware architecture (e.g. x86_64, arm, ppc, mips).
 	BootTime          time.Time `json:"boot_time"`               // Host boot time.
 	Containerized     *bool     `json:"containerized,omitempty"` // Is the process containerized.
-	Hostname          string    `json:"name"`                    // Hostname
+	Hostname          string    `json:"name"`                    // Hostname, lowercased.
 	IPs               []string  `json:"ip,omitempty"`            // List of all IPs.
 	KernelVersion     string    `json:"kernel_version"`          // Kernel version.
 	MACs              []string  `json:"mac"`                     // List of MAC addresses.
@@ -96,7 +101,7 @@ type OSInfo struct {
 // LoadAverage is the interface that wraps the LoadAverage method.
 // LoadAverage returns load info on the host
 type LoadAverage interface {
-	LoadAverage() LoadAverageInfo
+	LoadAverage() (*LoadAverageInfo, error)
 }
 
 // LoadAverageInfo contains load statistics
